requirements.txt
asgiref==3.5.2
cfgv==3.3.1
distlib==0.3.6
Django==4.1.2
django-cors-headers==3.13.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.12.2
identify==2.5.24
nodeenv==1.8.0
Pillow==9.2.0
platformdirs==3.5.3
pre-commit==3.3.3
psycopg2==2.9.5
PyJWT==2.6.0
pytz==2022.5
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.5
virtualenv==20.23.0
